Modularização de autenticação: adiciona cadastro.js para registro de usuários e login.js para autenticação com email e senha.
