+++ html/cadastro.html
@@ -62,8 +62,9 @@
     </section>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
     <script src="../scripts/main.js"></script>
-    <script src="../scripts/cadastro.js"></script>
-    <script type="module" src="../scripts/firebase.js"></script>
+    <script type="module" src="/scripts/firebase.js"></script>
+    <script type="module" src="/scripts/login.js"></script>
+    <script type="module" src="/scripts/cadastro.js"></script>
 </body>
 
 </html>
