+++ html/login.html
@@ -1,5 +1,6 @@
 <!DOCTYPE html>
 <html lang="pt-br">
+
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -8,13 +9,14 @@
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@100;400;700;900&display=swap" rel="stylesheet">
     <link rel="stylesheet" href="../styles/reset.css">
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
     <link rel="stylesheet" href="../styles/login.css">
 </head>
 
 <body>
     <header class="navbar navbar-expand-lg ">
-       
+
     </header>
     <section id="login-section">
         <div class="login-form">
@@ -22,11 +24,13 @@
             <form id="login-form">
                 <div class="mb-3">
                     <label for="email" class="form-label">E-mail</label>
-                    <input type="email" class="form-control" id="email" name="email" placeholder="Digite seu e-mail" required>
+                    <input type="email" class="form-control" id="email" name="email" placeholder="Digite seu e-mail"
+                        required>
                 </div>
                 <div class="mb-3">
                     <label for="password" class="form-label">Senha</label>
-                    <input type="password" class="form-control" id="password" name="password" placeholder="Digite sua senha" required>
+                    <input type="password" class="form-control" id="password" name="password"
+                        placeholder="Digite sua senha" required>
                 </div>
                 <button type="submit" class="btn btn-primary w-100">Entrar</button>
             </form>
@@ -54,7 +58,10 @@
 
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"></script>
     <script src="../scripts/main.js"></script>
-    <script src="../scripts/login.js"></script>
-    <script type="module" src="../scripts/firebase.js"></script>
+    <script type="module" src="/scripts/firebase.js"></script>
+    <script type="module" src="/scripts/login.js"></script>
+    <script type="module" src="/scripts/cadastro.js"></script>
+
 </body>
+
 </html>
