+++ scripts/firebase.js
@@ -1,6 +1,6 @@
 import { initializeApp } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-app.js";
-import { getAuth, createUserWithEmailAndPassword, signInWithEmailAndPassword } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-auth.js";
-import { getFirestore, collection, doc, setDoc, serverTimestamp } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-firestore.js";
+import { getAuth } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-auth.js";
+import { getFirestore } from "https://www.gstatic.com/firebasejs/11.0.1/firebase-firestore.js";
 
 const firebaseConfig = {
     apiKey: "AIzaSyCQXWDErfnbUaJAch6UJ0Vy65hEU60YWfY",
@@ -13,56 +13,9 @@ const firebaseConfig = {
 };
 
 const app = initializeApp(firebaseConfig);
+
+
 const auth = getAuth(app);
 const db = getFirestore(app);
 
-document.getElementById("login-form").addEventListener("submit", async (e) => {
-    e.preventDefault(); 
-
-    const email = document.getElementById('email').value;
-    const password = document.getElementById('password').value;
-    
-    try {
-        const userCredential = await signInWithEmailAndPassword(auth, email, password);
-        alert("Login realizado com sucesso!");
-    } catch (error) {
-        console.error("Erro ao logar: ", error.message);
-        alert("Erro ao logar: " + error.message);
-    }
-});
-
-
-document.getElementById("register-form").addEventListener("submit", async (e) => {
-    e.preventDefault();
-
-    const email = document.getElementById("email").value;
-    const password = document.getElementById("password").value;
-    const confirmPassword = document.getElementById("confirm-password").value;
-
-    if (password !== confirmPassword) {
-        alert("As senhas são diferentes.");
-        return;
-    }
-
-    const isOng = document.getElementById("isOng").checked;
-    const ongName = document.getElementById("nome-ong").value;
-    const cnpj = document.getElementById("cnpj").value;
-
-    try {
-        const userCredential = await createUserWithEmailAndPassword(auth, email, password);
-        const user = userCredential.user;
-
-        await setDoc(doc(db, "users", user.uid), {
-            email: email,
-            isOng: isOng,
-            ongName: isOng ? ongName : null,
-            cnpj: isOng ? cnpj : null,
-            createdAt: serverTimestamp()
-        });
-
-        alert("Usuário registrado com sucesso!");
-    } catch (error) {
-        console.error("Erro ao registrar usuário:", error.message);
-        alert("Erro ao registrar: " + error.message);
-    }
-});
+export { auth, db };
